Major refactor
merge ml pipeline with liana pipeline,
erase MetabMethodMeta
merge most of the function
include testing methods
get_ml_lr has to be merged
we still miss a few scores and estimations
