+++ liana/method/_Method.py
@@ -27,7 +27,8 @@ class MethodMeta:
                  specificity: str | None,
                  specificity_ascending: bool | None,
                  permute: bool,
-                 reference: str
+                 reference: str, 
+                 met: bool,
                  ):
         """
         Parameters
@@ -63,6 +64,7 @@ class MethodMeta:
         self.specificity = specificity
         self.specificity_ascending = specificity_ascending
         self.permute = permute
+        self.met = met
         self.reference = reference
 
     def describe(self):
@@ -161,7 +163,9 @@ class Method(MethodMeta):
                          specificity=_method.specificity,
                          specificity_ascending=_method.specificity_ascending,
                          permute=_method.permute,
-                         reference=_method.reference
+                         reference=_method.reference,
+                         met = _method.met
+
                          )
         self._method = _method
 
@@ -181,6 +185,7 @@ class Method(MethodMeta):
                  verbose: Optional[bool] = False,
                  n_perms: int = 1000,
                  seed: int = 1337,
+                 prop_missing_allowed: float = 0.98,
                  resource: Optional[DataFrame] = None,
                  inplace=True):
         """
@@ -249,6 +254,11 @@ class Method(MethodMeta):
                                resource=resource,
                                expr_prop=expr_prop,
                                min_cells=min_cells,
+                               met_est_resource_name=None,
+                               met_est_resource=None,
+                               est_fun=None,
+                               score_fun=None,
+                               prop_missing_allowed=prop_missing_allowed,
                                supp_columns=supp_columns,
                                return_all_lrs=return_all_lrs,
                                base=base,
@@ -267,3 +277,153 @@ class Method(MethodMeta):
 
 def _show_methods(methods):
     return concat([method.get_meta() for method in methods])
+
+
+class MetabMethod(MethodMeta):
+    """
+    liana's Method Class
+    """
+    def __init__(self, _method, **kwargs):
+        super().__init__(method_name=_method.method_name,
+                         reference=_method.reference,
+                         complex_cols=_method.complex_cols,
+                         add_cols=_method.add_cols,
+                         fun=_method.fun,
+                         magnitude=_method.magnitude,
+                         magnitude_ascending=_method.magnitude_ascending,
+                         specificity=_method.specificity,
+                         specificity_ascending=_method.specificity_ascending,
+                         permute=_method.permute,
+                         met=_method.met
+                         )
+        self._method = _method
+        
+
+
+    def __call__(self,
+                 adata: AnnData,
+                 groupby: str,
+                 resource: Optional[DataFrame] = None,
+                 resource_name: str = 'metalinksdb',
+                 met_est_resource_name: str = 'metalinksdb',
+                 met_est_resource: Optional[DataFrame] = None,
+                 est_fun: str = 'mean_per_cell',
+                 score_fun: str = 'cellphone',
+                 expr_prop: float = 0.1,
+                 min_cells: int = 5,
+                 base: float = 2.718281828459045,
+                 supp_columns: list = None,
+                 return_all_lrs: bool = False,
+                 prop_missing_allowed: float = 0.995,
+                 use_raw: Optional[bool] = True,
+                 layer: Optional[str] = None,
+                 verbose: Optional[bool] = False,
+                 n_perms: int = 1000,
+                 seed: int = 1337,
+                 inplace: bool = True,
+                 pass_mask: bool = True,
+                 est_only: bool = False,
+                 correct_fdr: bool = False,
+                 **kwargs):
+        """
+        Parameters
+        ----------
+        adata
+            Annotated data object.
+        groupby
+            The key of the observations grouping to consider.
+        output
+            Full MR calculation (CCC) or only metabolite estimation (ME).
+        resource_name
+            Name of the resource to be loaded and use for ligand-receptor inference.
+        met_est_resource_name
+            Name of the resource to be loaded and use for metabolite estimation.
+        met_est_resource
+            Metabolite-gene links to be used for metabolite estimation.
+        expr_prop
+            Minimum expression proportion for the ligands/receptors (and their subunits) in the
+            corresponding cell identities. Set to `0`, to return unfiltered results.
+        min_cells
+            Minimum cells per cell identity (`groupby`) to be considered for downstream analysis
+        base
+            Exponent base used to reverse the log-transformation of matrix. Note that this is
+            relevant only for the `logfc` method.
+        supp_columns
+            Any additional columns to be added from any of the methods implemented in
+            liana, or any of the columns returned by `scanpy.tl.rank_genes_groups`, each
+            starting with ligand_* or receptor_*. For example, `['ligand_pvals', 'receptor_pvals']`.
+            `None` by default.
+        return_all_lrs
+            Bool whether to return all LRs, or only those that surpass the `expr_prop`
+            threshold. Those interactions that do not pass the `expr_prop` threshold will
+            be assigned to the *worst* score of the ones that do. `False` by default.
+        use_raw
+            Use raw attribute of adata if present.
+        layer
+            Layer in anndata.AnnData.layers to use. If None, use anndata.AnnData.X.
+        de_method
+            Differential expression method. `scanpy.tl.rank_genes_groups` is used to rank genes
+            according to 1vsRest. The default method is 't-test'. Only relevant if p-values
+            are included in `supp_cols`
+        verbose
+            Verbosity flag
+        resource
+            Parameter to enable external resources to be passed. Expects a pandas dataframe
+            with [`ligand`, `receptor`] columns. None by default. If provided will overrule
+            the resource requested via `resource_name`
+        inplace
+            If true return `DataFrame` with results, else assign to `.layer`.
+
+        Returns
+        -------
+            If ``inplace = False``, returns a `DataFrame` with ligand-receptor results
+            Otherwise, modifies the ``adata`` object with the following key:
+            - :attr:`anndata.AnnData.uns` ``['liana_res']`` with the aforementioned DataFrame
+        """
+        if supp_columns is None:
+            supp_columns = []
+
+        ml_res = liana_pipe(adata=adata,
+                        groupby=groupby,
+                        resource_name=resource_name,
+                        resource=None,
+                        met_est_resource_name=met_est_resource_name,
+                        met_est_resource=None,
+                        expr_prop=expr_prop,
+                        est_fun=est_fun,
+                        score_fun=score_fun,
+                        min_cells=min_cells,
+                        supp_columns=supp_columns,
+                        return_all_lrs=return_all_lrs,
+                        base=base,
+                        de_method=None,
+                        verbose=verbose,
+                        use_raw=use_raw,
+                        n_perms=n_perms,
+                        seed=seed,
+                        layer=layer,
+                        _score = self._method, 
+                        pass_mask=pass_mask,
+                        est_only=est_only,
+                        correct_fdr=correct_fdr,
+                        prop_missing_allowed=prop_missing_allowed,
+                        **kwargs
+                        )
+        
+        if inplace:
+            if est_only:
+                if pass_mask:
+                    adata.obsm['metabolite_abundance'] = ml_res[0]
+                    adata.uns['mask'] = ml_res[1]
+                else:
+                    adata.obsm['metabolite_abundance'] = ml_res
+            else:
+                if pass_mask:
+                    adata.uns['CCC_res'] = ml_res[0]
+                    adata.obsm['metabolite_abundance'] = ml_res[1]
+                    adata.uns['mask'] = ml_res[2]
+                else:
+                    adata.uns['CCC_res'] = ml_res[0]
+                    adata.obsm['metabolite_abundance'] = ml_res[1]
+        
+        return None if inplace else ml_res
