+++ liana/method/__init__.py
@@ -1,6 +1,6 @@
-from ._Method import Method, MethodMeta, _show_methods
+from ._Method import Method, MethodMeta, MetabMethod, _show_methods
 from .sc._rank_aggregate import AggregateClass, _rank_aggregate_meta as aggregate_meta
-from .sc import cellphonedb, connectome, logfc, natmi, singlecellsignalr, geometric_mean, cellchat
+from .sc import cellphonedb, connectome, logfc, natmi, singlecellsignalr, geometric_mean, cellchat, metalinks
 
 
 import numpy as np
