--- liana/method/ml/scores/__init__.py
@@ -1,1 +1,0 @@
-from ._scores import _simple_prod, _product_score
