+++ liana/method/sc/__init__.py
@@ -1,9 +1,10 @@
 from ._rank_aggregate import _rank_aggregate_meta
-from ._cellphonedb import cellphonedb
-from ._logfc import logfc
-from ._natmi import natmi
-from ._connectome import connectome
-from ._singlecellsignalr import singlecellsignalr
-from ._geometric_mean import geometric_mean
-from ._cellchat import cellchat
+from ._cellphonedb import cellphonedb, _cpdb_score
+from ._logfc import logfc, _logfc_score
+from ._natmi import natmi, _natmi_score
+from ._connectome import connectome, _connectome_score
+from ._singlecellsignalr import singlecellsignalr, _sca_score
+from ._geometric_mean import geometric_mean, _gmean_score
+from ._cellchat import cellchat, _cellchat_score
+from ._metalinks import metalinks
