+++ liana/method/sc/_logfc.py
@@ -31,6 +31,7 @@ _logfc = MethodMeta(method_name="log2FC",
                     specificity='lr_logfc',
                     specificity_ascending=False,
                     permute=False,
+                    met = False,
                     reference='Dimitrov, D., Türei, D., Garrido-Rodriguez, M., Burmedi, P.L., '
                               'Nagai, J.S., Boys, C., Ramirez Flores, R.O., Kim, H., Szalai, B., '
                               'Costa, I.G. and Valdeolivas, A., 2022. Comparison of methods and '
