+++ liana/plotting/__init__.py
@@ -1,3 +1,3 @@
 from ._proximity_plot import proximity_plot
 from ._dotplot import dotplot, dotplot_by_sample
-from .ml._ml_dotplot import ml_dotplot, ml_dotplot_by_sample, gene_plot
+from .ml._ml_set_dotplot import set_dotplot
