+++ liana/tests/test_liana_pipe.py
@@ -18,6 +18,7 @@ complex_policy = 'min'
 key_cols = ['source', 'target', 'ligand_complex', 'receptor_complex']
 verbose = False
 base = 2.718281828459045
+prop_missing_allowed: float = 0.98,
 resource = None
 use_raw = True
 layer = None
@@ -33,6 +34,11 @@ def test_liana_pipe_defaults():
                               expr_prop=expr_prop,
                               min_cells=min_cells,
                               de_method=de_method,
+                              est_fun=None,
+                              score_fun=None,
+                              met_est_resource_name=None,
+                              met_est_resource=None,
+                              prop_missing_allowed=prop_missing_allowed,
                               base=base,
                               n_perms=n_perms,
                               seed=seed,
@@ -63,6 +69,11 @@ def test_liana_pipe_not_defaults():
                               min_cells=min_cells,
                               de_method='wilcoxon',
                               base=base,
+                              est_fun=None,
+                              score_fun=None,
+                              met_est_resource_name=None,
+                              met_est_resource=None,
+                              prop_missing_allowed=prop_missing_allowed,
                               n_perms=n_perms,
                               seed=seed,
                               verbose=verbose,
