+++ liana/tests/test_ml_estimations.py
@@ -1,97 +1,70 @@
-from liana.method._liana_pipe import liana_pipe, _expm1_base, _calc_log2fc
+import pandas
+from numpy import max, min
+from numpy.testing import assert_almost_equal
 
-import pathlib
-from scanpy.datasets import pbmc68k_reduced
-from pandas.testing import assert_frame_equal
-from pandas import read_csv
-import numpy as np
+from liana.method import metalinks
+    
+from liana.testing._toy_adata import get_toy_adata
 
-test_path = pathlib.Path(__file__).parent
-adata = pbmc68k_reduced()
+# load toy adata
+adata = get_toy_adata()
 
-groupby = 'bulk_labels'
-de_method = 't-test'
-resource_name = 'consensus'
-expr_prop = 0.1
-min_cells = 5
-complex_policy = 'min'
-key_cols = ['source', 'target', 'ligand_complex', 'receptor_complex']
-verbose = False
-base = 2.718281828459045
-resource = None
-use_raw = True
-layer = None
-n_perms = 5
-seed = 1337
+def test_ulm():
+    metalinks(adata, groupby='bulk_labels', est_fun = 'ulm', n_perms=10, est_only=True, pass_mask = False)
+    assert ('CCC_res' in adata.uns) == False
+    assert ('mask' in adata.uns) == False
+    assert adata.obsm['metabolite_abundance'].shape == (700, 29)
 
+def test_wsum():
+    metalinks(adata, groupby='bulk_labels', est_fun = 'wsum', n_perms=10, est_only=True, pass_mask = False)
+    assert ('CCC_res' in adata.uns) == False
+    assert ('mask' in adata.uns) == False
+    assert adata.obsm['metabolite_abundance'].shape == (700, 29)
 
-# Test ALL Default parameters
-def test_liana_pipe_defaults():
-    all_defaults = liana_pipe(adata=adata,
-                              groupby=groupby,
-                              resource_name=resource_name,
-                              expr_prop=expr_prop,
-                              min_cells=min_cells,
-                              de_method=de_method,
-                              base=base,
-                              n_perms=n_perms,
-                              seed=seed,
-                              verbose=verbose,
-                              _key_cols=key_cols,
-                              supp_columns=[],
-                              resource=resource,
-                              use_raw=use_raw,
-                              layer=layer
-                              )
+def test_wmean():
+    metalinks(adata, groupby='bulk_labels', est_fun = 'wmean', n_perms=10, est_only=True, pass_mask = False)
+    assert ('CCC_res' in adata.uns) == False
+    assert ('mask' in adata.uns) == False
+    assert adata.obsm['metabolite_abundance'].shape == (700, 29)
 
-    assert 1288 == all_defaults.shape[0]
-    assert 21 == all_defaults.shape[1]
-    assert 'prop_min' in all_defaults.columns
+def test_viper():
+    metalinks(adata, groupby='bulk_labels', est_fun = 'viper', n_perms=10, est_only=True, pass_mask = False)
+    assert ('CCC_res' in adata.uns) == False
+    assert ('mask' in adata.uns) == False
+    assert adata.obsm['metabolite_abundance'].shape == (700, 29)
 
-    exp_defaults = read_csv(test_path.joinpath("data/all_defaults.csv"), index_col=0)
-    exp_defaults.index = all_defaults.index
-    assert_frame_equal(all_defaults, exp_defaults, check_dtype=False,
-                       check_exact=False, check_index_type=False, rtol=1e-3)
+# TODO
 
+# def test_udt():
+#     metalinks(adata, groupby='bulk_labels', est_fun = 'udt', n_perms=10, est_only=True, pass_mask = False)
+#     assert ('CCC_res' in adata.uns) == False
+#     assert ('mask' in adata.uns) == False
+#     assert adata.obsm['metabolite_abundance'].shape == (700, 29)
 
-# Test NOT Default parameters
-def test_liana_pipe_not_defaults():
-    not_defaults = liana_pipe(adata=adata,
-                              groupby=groupby,
-                              resource_name=resource_name,
-                              expr_prop=0.2,
-                              min_cells=min_cells,
-                              de_method='wilcoxon',
-                              base=base,
-                              n_perms=n_perms,
-                              seed=seed,
-                              verbose=verbose,
-                              _key_cols=key_cols,
-                              supp_columns=['ligand_pvals', 'receptor_pvals'],
-                              resource=resource,
-                              use_raw=use_raw,
-                              layer=layer,
-                              return_all_lrs=True
-                              )
+# def test_mdt():
+#     metalinks(adata, groupby='bulk_labels', est_fun = 'mdt', n_perms=10, est_only=True, pass_mask = False)
+#     assert ('CCC_res' in adata.uns) == False
+#     assert ('mask' in adata.uns) == False
+#     assert adata.obsm['metabolite_abundance'].shape == (700, 29)
 
-    assert 4200 == not_defaults.shape[0]
-    assert 24 == not_defaults.shape[1]
-    assert all(np.isin(['lrs_to_keep'], not_defaults.columns))
-    assert all(np.isin(['ligand_pvals', 'receptor_pvals'], not_defaults.columns))
+# def test_ora():
+#     metalinks(adata, groupby='bulk_labels', est_fun = 'ora', n_perms=10, est_only=True, pass_mask = False)
+#     assert ('CCC_res' in adata.uns) == False
+#     assert ('mask' in adata.uns) == False
+#     assert adata.obsm['metabolite_abundance'].shape == (700, 29)
 
-    exp_defaults = read_csv(test_path.joinpath("data/not_defaults.csv"), index_col=0)
-    exp_defaults.index = not_defaults.index
-    assert_frame_equal(not_defaults, exp_defaults, check_dtype=False,
-                       check_index_type=False, check_exact=False, rtol=1e-3)
+# def test_gsea():
+#     metalinks(adata, groupby='bulk_labels', est_fun = 'gsea', n_perms=10, est_only=True, pass_mask = False)
+#     assert ('CCC_res' in adata.uns) == False
+#     assert ('mask' in adata.uns) == False
+#     assert adata.obsm['metabolite_abundance'].shape == (700, 29)
+
+# def test_gsva():
+#     metalinks(adata, groupby='bulk_labels', est_fun = 'gsva', n_perms=10, est_only=True, pass_mask = False)
+#     assert ('CCC_res' in adata.uns) == False
+#     assert ('mask' in adata.uns) == False
+#     assert adata.obsm['metabolite_abundance'].shape == (700, 29)
 
 
-def test_expm1_fun():
-    expm1_mat = _expm1_base(base, adata.raw.X.data)
-    np.testing.assert_almost_equal(np.sum(expm1_mat), 1057526.4, decimal=1)
 
-
-def test_calc_log2fc():
-    adata.layers['normcounts'] = adata.raw.X.copy()
-    adata.layers['normcounts'].data = _expm1_base(base, adata.raw.X.data)
-    adata.obs['label'] = adata.obs.bulk_labels
-    np.testing.assert_almost_equal(np.mean(_calc_log2fc(adata, "Dendritic")), -0.123781264)
+
